requirements.txt
Django>=4.2
python-jobspy
pdfplumber
groq
requests
beautifulsoup4
python-dotenv
